:sparkles: GET /api/readme
