+++ src/script/server/api/interfaces/index.js
@@ -1,0 +1,6 @@
+const connect = require('./connect')
+const readme = require('./readme')
+
+module.exports = (app, basePath) => ({
+  readme: connect(app, basePath)(readme),
+})
