+++ src/script/server/api/interfaces/readme.js
@@ -1,0 +1,11 @@
+const doc = require('../services/doc')
+
+const get = {
+  method: 'get',
+  path: 'readme',
+  handler: async () => doc.generate('markdown'),
+}
+
+module.exports = {
+  get,
+}
