+++ src/script/server/api/services/doc/index.js
@@ -1,4 +1,4 @@
 const { connectAll } = require('../../../../redux')
 const service = require('./service')
 
-module.exports = { NAME: 'DOC', ...connectAll(service) }
+module.exports = connectAll(service)
